fix: 보고서 생성 시 마지막 날짜 지정 오류 수정

See merge request s11-final/S11P31C106!77

+++ backend/src/main/java/ssafy/modo/jamkkaebi/domain/manager/dto/response/ReportResponseDto.java
@@ -33,7 +33,7 @@ public class ReportResponseDto {
     @Data
     @Builder
     public static class WeekData {
-        private List<Integer> lastWeek;
-        private List<Integer> thisWeek;
+        private List<Long> lastWeek;
+        private List<Long> thisWeek;
     }
 }
